refactored splitter components to be more modular and updated test cases to match the refactoring
