+++ .pre-commit-config.yaml
@@ -8,7 +8,14 @@ repos:
     rev: v0.4.2
     hooks:
       - id: ruff
-        args: ["--fix", "--extend-ignore=E402", "--extend-ignore=F841"]
+        args:
+          [
+            "--fix",
+            "--extend-ignore=E402",
+            "--extend-ignore=F841",
+            "--extend-ignore=F403",
+            "--extend-ignore=F405",
+          ]
   - repo: https://github.com/pre-commit/pre-commit-hooks
     rev: v4.0.1
     hooks:
