+++ README.md
@@ -32,16 +32,141 @@ This project uses the UV package manager. To install missing_text, follow these 
 
 ## Usage
 
-Here's a quick example of how to use missing_text:
+### PDF Processing
+
+The package provides powerful PDF processing capabilities with both synchronous and asynchronous options:
 
 ```python
+from missing_text import sync_extract_pdf, async_extract_pdf, extract_pdfs
 
-from missing_text.hello_missing import hello_missing
-result = hello_missing()
-print(result)
+# Process a single PDF synchronously
+with open("document.pdf", "rb") as f:
+    content = sync_extract_pdf(
+        f.read(),
+        text=True,      # Extract text
+        table=True,     # Extract tables
+        image=True,     # Extract images with OCR
+        encode_page=True, # Get base64 encoded page images
+        segment=True    # Get content segments with bounding boxes
+    )
+
+# Process a PDF asynchronously
+async def process_pdf():
+    content = await async_extract_pdf("path/to/pdf", safe_mode=True)
+    print(content["text"])  # Access extracted text
+    print(content["tables"]) # Access extracted tables
+    print(content["images"]) # Access extracted images with OCR
+    print(content["pages"])  # Access page images
+    print(content["segments"]) # Access content segments
+
+# Process multiple PDFs from a directory
+results = extract_pdfs(
+    "path/to/pdf_directory",
+    safe_mode=True,
+    text=True,
+    table=True,
+    image=True
+)
 ```
 
-For more detailed usage instructions, please refer to the [documentation](docs/README.md).
+### Text Splitting
+
+The package includes various text splitting utilities for LLM context preparation:
+
+```python
+from missing_text.splitter import (
+    character_splitter,
+    sentence_splitter,
+    paragraph_splitter,
+    markdown_header_splitter,
+    json_key_splitter,
+    html_tag_attribute_splitter,
+    latex_section_splitter,
+    recursive_character_splitter,
+)
+
+# Split text by characters with overlap
+chunks = character_splitter(
+    "Your long text here",
+    chunk_size=100,
+    overlap=20
+)
+
+# Split by sentences
+sentences = sentence_splitter("Multiple sentences. Like this one. And this.")
+
+# Split markdown by headers
+sections = markdown_header_splitter(
+    "# Section 1\nContent\n# Section 2\nMore content",
+    level=1
+)
+
+# Advanced recursive splitting with overlap
+chunks = recursive_character_splitter(
+    text="Your long document text here",
+    character_size=800,
+    overlap=100,
+    delimiters=["\n\n", "\n", "[.!?]", ",", " ", ""]
+)
+```
+
+### Text Embeddings
+
+The package supports multiple embedding options:
+
+```python
+from missing_text.embed.sentence_transformers import sentence_transformer_embedder
+
+# Generate embeddings using Sentence Transformers
+texts = ["First sentence", "Second sentence", "Third sentence"]
+embeddings = sentence_transformer_embedder(
+    texts,
+    model_name="all-MiniLM-L6-v2"
+)
+```
+
+### FastAPI Integration
+
+The package includes a FastAPI server for document processing:
+
+```python
+# Start the FastAPI server
+missing fastapi --host 0.0.0.0 --port 8000
+
+# Use the API endpoints
+curl -X POST "http://localhost:8000/extract/pdf" \
+     -H "accept: application/json" \
+     -H "Content-Type: multipart/form-data" \
+     -F "file=@document.pdf"
+
+# Process PDF from bytes
+curl -X POST "http://localhost:8000/extract/pdf-bytes" \
+     -H "Content-Type: application/pdf" \
+     --data-binary @document.pdf
+
+# Process PDF from path
+curl -X POST "http://localhost:8000/extract/pdf-path" \
+     -H "Content-Type: application/json" \
+     -d '{"file_path": "/path/to/document.pdf"}'
+```
+
+### Streamlit UI
+
+The package includes a Streamlit interface for visual document processing:
+
+```bash
+# Start the Streamlit app
+missing streamlit --host localhost --port 8501
+```
+
+The Streamlit UI provides:
+
+- PDF upload and processing
+- Text extraction visualization
+- Table extraction with preview
+- Image extraction with OCR results
+- Content segmentation visualization
+- JSON export of extracted content
 
 ## CLI Usage
