+++ docs/docs/code-example.md
@@ -1,0 +1,5 @@
+```python
+# Example code
+def hello_world():
+    print("Hello, world!")
+```
